requirements.txt
# Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Visualisation et interface
pygame>=2.5.0

# Utilitaires
progressbar2>=4.2.0
openpyxl>=3.1.0
